backend/requirements.txt
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
motor>=3.3.0,<4.0.0
pydantic-settings>=2.1.0,<3.0.0
slowapi>=0.1.9,<1.0.0
redis>=5.0.0,<6.0.0
secure>=0.3.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
httpx>=0.27.0,<1.0.0
